+++ SistemaAlzeimer/Lista.h
@@ -18,7 +18,7 @@ public:
     Lista();
     void inserir(Paciente* paciente);
     void deletar(string nome);
-    int contador();
+    float contador();
     int alterar(string nome);
     void imprimir();
     void consultar(string nome);
